Make Item molecules and change interface place
